Add files via upload
Main is functional, the rest is for white box testing.

A handful of errors, the undo function doesn't undo the existence of a block feature. Don't know what's happening to values under a block position when an undo occurs. It's likely that they simply stop being acted on but are still part of the map.

Would like to improve speed and add animation.

Have the option to move blocks across the screen, save and load games.